+++ tests/api_infinity_chess/RP_lista_tutores_por_sede/test_RPL008.py
@@ -2,10 +2,10 @@ import requests
 import pytest
 
 @pytest.mark.smoke
-def test_RPL008_obtenerTutorPorSedeValida (getUrl):
+def test_RPL008_obtener_tutor_por_sede_valida (get_url):
     endpoint = "obtenerTutores/Modulo 4"
-    lista_url = getUrl + endpoint
+    lista_url = get_url + endpoint
     response = requests.get(lista_url)
     assert response.status_code == 200
-    listaTutores = response.json()
-    assert len (listaTutores) > 0
+    lista_tutores = response.json()
+    assert len (lista_tutores) > 0
