--- tests/api_infinitychess/cambiarEstadoDeTutor/SSL011.py
@@ -1,17 +1,0 @@
-import requests
-import pytest
-import random
-from src.api_infinityChess.obtenerTutores import obtenerTutoresActivos
-
-@pytest.mark.smoke
-def test_validaciónDeValorInválidoEnCampoEstado (getUrl):
-     listaTutores = obtenerTutoresActivos(getUrl)
-     CODTUTOR = random.choice(listaTutores)["CODTUTOR"]
-     print(f"Modificando estado del tutor: {CODTUTOR}")
-     endpoint = "actualizarEstadoTutor/" + CODTUTOR
-     payload = {
-          "ESTADO" : ""
-     }
-     urlFinal = getUrl + endpoint
-     response = requests.put(urlFinal, json=payload)
-     assert response.status_code == 400
