--- tests/api_infinitychess/cambiarEstadoDeTutor/SSL012.py
@@ -1,14 +1,0 @@
-import requests
-import pytest
-import random
-from src.api_infinityChess.obtenerTutores import obtenerTutoresActivos
-
-@pytest.mark.smoke
-def test_solicitudSinBody (getUrl):
-     listaTutores = obtenerTutoresActivos(getUrl)
-     CODTUTOR = random.choice(listaTutores)["CODTUTOR"]
-     print(f"Modificando estado del tutor: {CODTUTOR}")
-     endpoint = "actualizarEstadoTutor/" + CODTUTOR
-     urlFinal = getUrl + endpoint
-     response = requests.put(urlFinal)
-     assert response.status_code == 400
