--- tests/api_infinitychess/obtenerListaEstudiantes/FML001.py
@@ -1,9 +1,0 @@
-import requests
-import pytest
-
-@pytest.mark.smoke
-def test_validarSEDEvacíaRetornaListaVacía(getUrl):
-    endpoint = "obtenerEstudiantes/"
-    lista_url = getUrl + endpoint
-    response = requests.get(lista_url)
-    assert response.status_code == 404
