--- tests/api_infinitychess/obtenerTrabajadores/SVT001.py
@@ -1,16 +1,0 @@
-import requests
-import pytest
-import random
-from src.api_infinityChess.obtenerTrabajadores import obtenerTrabajadores
-
-@pytest.mark.smoke
-def test_obtenerTrabajadorExistentePorIdValido(getUrl):
-    listaTrabajdores = obtenerTrabajadores(getUrl)
-    CODTRABAJADOR = random.choice(listaTrabajdores)["CODTRABAJADOR"]
-    print(f"Trabajador escogido es: {CODTRABAJADOR}")
-    endpoint = "obtenerTrabajadores/" + CODTRABAJADOR
-    lista_url = getUrl + endpoint
-    response = requests.get(lista_url)
-    assert response.status_code == 200
-    print (response.json[0].get("NOMBRETRABAJADOR")) 
-
