--- tests/api_infinitychess/obtenerTutoresActivos/SSL006.py
@@ -1,14 +1,0 @@
-import requests
-import pytest
-
-@pytest.mark.smoke
-def test_solicitudConHeaders(getUrl):
-    endpoint = "obtenerTutoresActivos"
-    lista_url = getUrl + endpoint
-    headers = {
-    }
-    response = requests.get(lista_url, headers=headers)
-    assert response.status_code == 200
-    listaTutores = response.json()
-    for tutor in listaTutores:
-        assert tutor.get("ESTADO") == "Activo", f"Tutor inactivo encontrado: {tutor}"
