--- tests/api_infinitychess/verificarLaMateria/MOCM002.py
@@ -1,15 +1,0 @@
-import requests
-import pytest
-import random
-from src.api_infinityChess.obtenerCurso import obtenerCursos
-
-@pytest.mark.smoke
-def test_ValidarQueSeRecuperenLaMateriaConNombreDeCurso(getUrl):
-    listaMaterias = obtenerCursos(getUrl)
-    CURSO = random.choice(listaMaterias)["CURSO"]
-    print(f"Materia escogida es: {CURSO}")
-    endpoint = "verificarCurso/" + CURSO
-    lista_url = getUrl + endpoint
-    response = requests.get(lista_url)
-    assert response.status_code == 200
-
